Addressed most PR feedback
