+++ composer/utils/__init__.py
@@ -1,20 +1,13 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 """Helper utilities."""
+from composer.utils.checkpoint import load_checkpoint, save_checkpoint
 from composer.utils.file_retriever import get_file
-from composer.utils.iter_helpers import ensure_tuple as ensure_tuple
-from composer.utils.iter_helpers import iterate_with_pbar as iterate_with_pbar
-from composer.utils.iter_helpers import map_collection as map_collection
-from composer.utils.object_store import ObjectStoreProvider as ObjectStoreProvider
-from composer.utils.object_store import ObjectStoreProviderHparams as ObjectStoreProviderHparams
-from composer.utils.string_enum import StringEnum as StringEnum
+from composer.utils.iter_helpers import ensure_tuple, iterate_with_pbar, map_collection
+from composer.utils.object_store import ObjectStoreProvider, ObjectStoreProviderHparams
+from composer.utils.string_enum import StringEnum
 
 __all__ = [
-    'ensure_tuple',
-    'iterate_with_pbar',
-    'map_collection',
-    'get_file',
-    'ObjectStoreProvider',
-    'ObjectStoreProviderHparams',
-    'StringEnum',
+    'ensure_tuple', 'iterate_with_pbar', 'map_collection', 'get_file', 'ObjectStoreProvider',
+    'ObjectStoreProviderHparams', 'StringEnum', "load_checkpoint", "save_checkpoint"
 ]
